Fixed error in request to re-post a post and importations
