+++ src/App.js
@@ -3,9 +3,8 @@ import { BrowserRouter, Route, Routes} from "react-router-dom";
 import UserContext from "./contexts/UserContext";
 import SignUp from "./Components/Login/SignUp";
 import SignIn from "./Components/Login/SignIn";
-import InputSearchUsers from "./Components/InputSearch/InputSearchUsers";
 import TelaUser from "./Pages/TelaUser.js";
-import PostsPage from "./Pages/PostsPage";
+import PostsPage from "./Components/Posts/PostsPage";
 import HashtagPage from "./Pages/HashtagPage";
 import UrlContext from "./contexts/UrlContext";
 import SearchNewUpdates from "./Components/Posts/SearchNewUpdates";
